+++ exporters/stdout/stdoutlog/config.go
@@ -38,7 +38,6 @@ func newConfig(options []Option) config {
 	for _, opt := range options {
 		cfg = opt.apply(cfg)
 	}
-
 	return cfg
 }